[2024] day 2
part 1 was easy, in part 2 i was making the mistake
of removing the element that causes the first
violation instea. that would not be correct
in some cases since the request is to check whether
it becomes safe after removing any one element
so i ended up trying them all.
I don't know if it is possible to do this online
while checking for violation, but that would be
surely better since i could do the check in one
sweep