[2024] day 3
i used ELL parsing, ignore all characters except
the ones that can be the start of a valid substring
and have a procedure that checks if it is correct
one character at a time.
This is of course overkill and i could have solved
this with regexes faster since the language is
regular and not context-free.
Took about 45 minutes to code.