[2024] day 11
part 1 is easy since one can just simulate all
the rules on an actual list, for part 2 you need
a more creative approach since the list will never
fit in memory after some iterations.
The insight was noticing that given a stone
the result is always the same so i can just count
how many there are and keep track of that.
This way i can keep track of just 3767 numbers
in a dict instead of 200000 billion, which is
good for performance