[2024] day 14
the solution to part 2 is funny, i think it can
be solved by some weird modular maths but i
think the principle i used is at least sound.
The problem was not constructed in a way that i
actually needed to track if it was a tree because
they never cluster in other shapes, except on
the two axis separately.